+++ src/app/(site)/_components/list-item.tsx
@@ -9,7 +9,7 @@ interface ListItemProps {
   href: string;
 }
 
-const ListItem: React.FC<ListItemProps> = ({ image, name, href }) => {
+export const ListItem: React.FC<ListItemProps> = ({ image, name, href }) => {
   const router = useRouter();
 
   const onClick = () => {
@@ -19,17 +19,15 @@ const ListItem: React.FC<ListItemProps> = ({ image, name, href }) => {
   return (
     <button
       onClick={onClick}
-      className="relative group flex items-center rounded-md overflow-hidden gap-x-4 bg-neutral-100/10 hover:bg-neutral-100/20 transition pr-4"
+      className="relative flex items-center pr-4 overflow-hidden transition rounded-md group gap-x-4 bg-neutral-100/10 hover:bg-neutral-100/20"
     >
       <div className="relative min-h-[64px] min-w-[64px]">
         <Image className="object-cover" fill src={image} alt="Image" />
       </div>
-      <p className="font-medium truncate py-5">{name}</p>
-      <div className="absolute transition opacity-0 rounded-full flex items-center justify-center bg-blue-500 p-4 drop-shadow-md right-5 group-hover:opacity-100 hover:scale-110">
+      <p className="py-5 font-medium truncate">{name}</p>
+      <div className="absolute flex items-center justify-center p-4 transition bg-blue-500 rounded-full opacity-0 drop-shadow-md right-5 group-hover:opacity-100 hover:scale-110">
         <FaPlay className="text-black" />
       </div>
     </button>
   );
 };
-
-export default ListItem;
