+++ src/app/(site)/_components/play-button.tsx
@@ -1,27 +1,9 @@
 import { FaPlay } from 'react-icons/fa';
 
-const PlayButton = () => {
+export const PlayButton = () => {
   return (
-    <button
-      className="
-        transition
-        opacity-0
-        rounded-full
-        flex
-        items-center
-        bg-blue-400
-        p-4
-        drop-shadow-md
-        translate
-        translate-y-1/4
-        group-hover:opacity-100
-        group-hover:translate-y-0
-        hover:scale-110
-        "
-    >
+    <button className="flex items-center p-4 transition bg-blue-400 rounded-full opacity-0  drop-shadow-md translate translate-y-1/4 group-hover:opacity-100 group-hover:translate-y-0 hover:scale-110">
       <FaPlay className="text-black" />
     </button>
   );
 };
-
-export default PlayButton;
